Add new deferred node reader function
Instead of eagerly outputting a vector of links, we can now do (almost)
no work until a link is actually asked for, which can speed things up
considerably when looking for a single link.

The handling of the single-child node special case is a bit dumb and
could probably be improved somehow, but the overall idea is already
helping quite a lot.