Redo API again to include only two types
Now there are only two types, one for representing a node in the trie
and one for representing a link, and they both carry the pointer to the
buffer with them, so they're independent and don't need a separate
full-trie object around. A node representing the root replaces the trie
type.