Add a "thin node" API
`Node` contains a reference to the underlying trie buffer; that allows
for a nice and convenient API, but there are use cases where reducing
the size of the struct has value (e.g., storing a bunch of nodes in a
priority queue, which involves shuffling them around in memory a whole
lot). This introduces a `ThinNode` that omits the buffer; using it for
appropriate applications allows a noticeable improvement in performance.