Further reduce node size by storing location as a non-Option
Using `Option<usize>` for the location adds a full eight bytes to the
struct size over just `usize`, so switch to the latter and use
`usize::MAX` in place of `None`; this can have a noticeable impact on
performance. Slightly sketchy from a code hygiene perspective, but it's
wrapped up pretty well and has no effect on the public API, so I'm okay
with it. I tried using `Option<NonZeroUsize>`, since zero isn't a valid
offset, but that turns out to be slower, even with unsafe unchecked
creation. Maybe I was just using it wrong, but I gave up.

Along with the thin node API, the reduction in size can give speedups of
a few percent up to almost 20%.